clear mappings for table, move reusable mapping functions to own components
