+++ src/components/Manager/MappingsFunctions/MappingSearch.jsx
@@ -1,21 +1,18 @@
 import { Checkbox, Form, Tooltip } from 'antd';
 import { useContext, useEffect, useRef, useState } from 'react';
 import { myContext } from '../../../App';
-import {
-  ellipsisString,
-  ontologyFilter,
-  ontologyReducer,
-  systemsMatch,
-} from '../../Manager/Utilitiy';
-import { ModalSpinner } from '../../Manager/Spinner';
+import { ellipsisString, ontologyReducer, systemsMatch } from '../Utilitiy';
+import { ModalSpinner } from '../Spinner';
+import { MappingContext } from '../../../MappingContext';
 
-export const MappingReset = ({
-  searchProp,
+export const MappingSearch = ({
+  editMappings,
   setEditMappings,
   form,
   mappingsForSearch,
   reset,
   onClose,
+  searchProp,
 }) => {
   const { searchUrl } = useContext(myContext);
   const [page, setPage] = useState(0);
@@ -27,16 +24,27 @@ export const MappingReset = ({
   const [lastCount, setLastCount] = useState(0); //save last count as count of the results before you fetch data again
   const [filteredResultsCount, setFilteredResultsCount] = useState(0);
 
+  const { setExistingMappings, setFilteredMappings } =
+    useContext(MappingContext);
+
   let ref = useRef();
+
+  const onExistingChange = checkedValues => {
+    setExistingMappings(checkedValues);
+  };
+
+  const onFilteredChange = checkedvalues => {
+    setFilteredMappings(checkedvalues);
+  };
 
   // since the code is passed through editMappings, the '!!' forces it to be evaluated as a boolean.
   // if there is a code being passed, it evaluates to true and runs the search function.
   // The function is run when the page changes and when the code changes.
   useEffect(() => {
-    if (!!searchProp) {
+    if (!!editMappings) {
       fetchResults(page);
     }
-  }, [page, searchProp]);
+  }, [page, editMappings]);
 
   /* Pagination is handled via a "View More" link at the bottom of the page. 
   Each click on the "View More" link makes an API call to fetch the next 15 results.
@@ -61,7 +69,7 @@ export const MappingReset = ({
 
   // The function that makes the API call to search for the passed code.
   const fetchResults = page => {
-    if (!!!searchProp) {
+    if (!!!editMappings) {
       return undefined;
     }
     setLoading(true);
@@ -116,9 +124,22 @@ export const MappingReset = ({
     setPage(page + 1);
   };
 
+  const viewMorePagination = (
+    <span
+      className="view_more_link"
+      onClick={e => {
+        handleViewMore(e);
+        // the lastCount being set to resultsCount prior to fetching the next batch of results
+        setLastCount(resultsCount);
+      }}
+    >
+      View More
+    </span>
+  );
+
   // The display for the checkboxes. The index is set to the count of the results before you fetch the new batch of results
   // again + 1, to move the scrollbar to the first result of the new batch.
-  const checkBoxDisplay = (d, index) => {
+  const newSearchDisplay = (d, index) => {
     index === lastCount + 1;
     return (
       <>
@@ -140,25 +161,70 @@ export const MappingReset = ({
                 </a>
               </div>
             </div>
-            <div>{ellipsisString(d?.description[0], '100')}</div>
+            <div>{ellipsisString(d?.description?.[0], '100')}</div>
           </div>
         </div>
       </>
     );
   };
 
-  const viewMorePagination = (
-    <span
-      className="view_more_link"
-      onClick={e => {
-        handleViewMore(e);
-        // the lastcount being set to resultsCount prior to fetching the next batch of results
-        setLastCount(resultsCount);
-      }}
-    >
-      View More
-    </span>
-  );
+  // Display for existing mappings.
+  const existingMappingDisplay = (d, index) => {
+    return (
+      <>
+        <div
+          key={index}
+          // prettier-ignore
+          className="modal_search_result"
+        >
+          <div>
+            <div className="modal_term_ontology">
+              <div>
+                <b>{d.display}</b>
+              </div>
+              <div>
+                {/* <a href={d.iri} target="_blank"> */}
+                {d.code}
+                {/* </a> */}
+              </div>
+            </div>
+            <div>{ellipsisString(d?.description?.[0], '100')}</div>
+          </div>
+        </div>
+      </>
+    );
+  };
+
+  // Iteratesw through the array of previously selected mappings. Returns a JSON stringified object that is pushed to a separate array.
+  // That array is returned to use as default checked values separate from the search results.
+  const initialChecked = () => {
+    let initialMappings = [];
+
+    mappingsForSearch.forEach((m, index) => {
+      const val = JSON.stringify({
+        code: m?.code,
+        display: m?.display,
+        // description: m.description[0],
+        system: m?.system,
+      });
+      initialMappings.push(val);
+    });
+    return initialMappings;
+  };
+
+  // Sets existingMappings to the mappings that have already been mapped to pass them to the body of the PUT call on save.
+  useEffect(() => {
+    setExistingMappings(mappingsForSearch);
+  }, []);
+
+  // Creates a Set that excludes the mappings that have already been selected.
+  // Then filteres the existing mappings out of the results to only display results that have not yet been selected.
+  const filteredResults = () => {
+    const codesToExclude = new Set(mappingsForSearch.map(m => m.code));
+    const updatedResults = results.filter(r => !codesToExclude.has(r.obo_id));
+    return updatedResults;
+  };
+  const filteredResultsArray = filteredResults();
 
   return (
     <>
@@ -168,39 +234,69 @@ export const MappingReset = ({
             <>
               <div className="modal_search_results">
                 <div className="modal_search_results_header">
-                  <h3>Search results for: {searchProp}</h3>
+                  <h3>Search results for: {editMappings?.code}</h3>
                 </div>
                 {/* ant.design form displaying the checkboxes with the search results.  */}
                 {results?.length > 0 ? (
                   <div className="result_container">
                     <Form form={form} layout="vertical">
                       <Form.Item
-                        name={['mappings']}
+                        // If it is NOT in reset state (i.e. edit mode), default values are checked. If reset is set, default values are blank, since all values were deleted.
+                        initialValue={!reset ? initialChecked() : null}
+                        name={['existing_mappings']}
                         valuePropName="value"
                         rules={[
                           {
-                            required: true,
-                            message: 'Please make a selection.',
+                            required: false,
                           },
                         ]}
                       >
-                        {results?.length > 0 ? (
+                        {mappingsForSearch?.length > 0 && (
                           <Checkbox.Group
                             className="mappings_checkbox"
-                            options={results?.map((d, index) => {
+                            options={mappingsForSearch?.map((d, index) => {
+                              return {
+                                value: JSON.stringify({
+                                  code: d.code,
+                                  display: d.display,
+                                  // description: d.description[0],
+                                  system: d.system,
+                                }),
+                                label: existingMappingDisplay(d, index),
+                              };
+                            })}
+                            onChange={onExistingChange}
+                          />
+                        )}
+                      </Form.Item>
+
+                      <Form.Item
+                        name={['filtered_mappings']}
+                        valuePropName="value"
+                        rules={[
+                          {
+                            required: false,
+                          },
+                        ]}
+                      >
+                        {filteredResultsArray?.length > 0 && (
+                          <Checkbox.Group
+                            className="mappings_checkbox"
+                            options={filteredResultsArray?.map((d, index) => {
                               return {
                                 value: JSON.stringify({
                                   code: d.obo_id,
                                   display: d.label,
                                   // description: d.description[0],
-                                  system: systemsMatch(d?.obo_id.split(':')[0]),
+                                  system: systemsMatch(
+                                    d?.obo_id?.split(':')[0]
+                                  ),
                                 }),
-                                label: checkBoxDisplay(d, index),
+                                label: newSearchDisplay(d, index),
                               };
                             })}
+                            onChange={onFilteredChange}
                           />
-                        ) : (
-                          ''
                         )}
                       </Form.Item>
                     </Form>
