+++ src/components/Manager/MappingsFunctions/MappingsFunctions.scss
@@ -1,0 +1,4 @@
+.reset-modal,
+.clear-mappings {
+  top: 25vh;
+}
